requirements/requirements-install-requires.txt
djangorestframework>=3.11.2
python-dateutil>=2.8.1
